+++ CHANGELOG.md
@@ -10,6 +10,13 @@
 
 **Note**: Gaps between patch versions are faulty/broken releases.
 
+## v2.1.0
+
+- **New Feature**
+  - upgrade to tcomb v2.2
+    - validation of `intersection` types
+    - error messages are formatted accordingly
+
 ## v2.0.1
 
 - **Internal**